sqlite-mcp-client-py/requirements.txt
requests
sseclient-py 